tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/toDataURL.unrecognised.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: toDataURL.unrecognised</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>toDataURL.unrecognised</h1>
<p class="desc">toDataURL with an unhandled type returns a PNG</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("toDataURL with an unhandled type returns a PNG");
_addTest(function(canvas, ctx) {

var data = canvas.toDataURL('image/example');
assert_regexp_match(data, /^data:image\/png[;,]/);


});
</script>
